🔧 fix(App.tsx): remove unused imports and variables for better code cleanliness ✨ feat(authContext.tsx): move autoLogin logic from App component to AuthProvider to improve separation of concerns 🔧 fix(index.tsx): reorder context providers to ensure correct execution order 🔧 fix(tabsContext.tsx): add comment to clarify the purpose of the useEffect hook 🔧 fix(typesContext.tsx): add comment to clarify the purpose of the useEffect hook 🔧 fix(routes.tsx): rename ProtectedAdminRoute to ProtectedRoute for consistency
