+++ src/frontend/src/contexts/index.tsx
@@ -1,6 +1,8 @@
 import { ReactNode } from "react";
+import { BrowserRouter } from "react-router-dom";
 import { ReactFlowProvider } from "reactflow";
 import { TooltipProvider } from "../components/ui/tooltip";
+import { ApiInterceptor } from "../controllers/API/api";
 import { SSEProvider } from "./SSEContext";
 import { AlertProvider } from "./alertContext";
 import { AuthProvider } from "./authContext";
@@ -9,34 +11,32 @@ import { LocationProvider } from "./locationContext";
 import { TabsProvider } from "./tabsContext";
 import { TypesProvider } from "./typesContext";
 import { UndoRedoProvider } from "./undoRedoContext";
-import { BrowserRouter } from "react-router-dom";
-import { ApiInterceptor } from "../controllers/API/api";
 
 export default function ContextWrapper({ children }: { children: ReactNode }) {
   //element to wrap all context
   return (
     <>
-    <BrowserRouter>
-      <AuthProvider>
-        <TooltipProvider>
-          <ReactFlowProvider>
-            <DarkProvider>
-              <TypesProvider>
-                <LocationProvider>
-                  <AlertProvider>
-                  <ApiInterceptor />
-                    <SSEProvider>
-                      <TabsProvider>
-                        <UndoRedoProvider>{children}</UndoRedoProvider>
-                      </TabsProvider>
-                    </SSEProvider>
-                  </AlertProvider>
-                </LocationProvider>
-              </TypesProvider>
-            </DarkProvider>
-          </ReactFlowProvider>
-        </TooltipProvider>
-      </AuthProvider>
+      <BrowserRouter>
+        <AlertProvider>
+          <AuthProvider>
+            <TooltipProvider>
+              <ReactFlowProvider>
+                <DarkProvider>
+                  <TypesProvider>
+                    <LocationProvider>
+                      <ApiInterceptor />
+                      <SSEProvider>
+                        <TabsProvider>
+                          <UndoRedoProvider>{children}</UndoRedoProvider>
+                        </TabsProvider>
+                      </SSEProvider>
+                    </LocationProvider>
+                  </TypesProvider>
+                </DarkProvider>
+              </ReactFlowProvider>
+            </TooltipProvider>
+          </AuthProvider>
+        </AlertProvider>
       </BrowserRouter>
     </>
   );
