+++ src/frontend/src/contexts/tabsContext.tsx
@@ -129,6 +129,7 @@ export function TabsProvider({ children }: { children: ReactNode }) {
   }
 
   useEffect(() => {
+    // If the user is authenticated, fetch the types. This code is important to check if the user is auth because of the execution order of the useEffect hooks.
     if (getAuthentication() === true) {
       // get data from db
       //get tabs locally saved
